+++ tests/cc/usualconv.c
@@ -11,7 +11,7 @@ static void test_usual_conv() {
     expect(4, sizeof((char)0 + (char)0));
     expect(4, sizeof((char)0 + (bool)0));
     expect(4, sizeof((char)0 + (int)0));
-    expect(8, sizeof((char)0 + (long)0));
+    expect(8, sizeof((char)0 + (long long)0));
     expect(8, sizeof((char)0 + (double)0));
 }
